\name{BayesPref-package}
\alias{BayesPref-package}
\alias{BayesPref}
\docType{package}
\title{
\packageTitle{BayesPref}
}
\description{
\packageDescription{BayesPref}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesPref}
\packageIndices{BayesPref}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BayesPref}

Maintainer: \packageMaintainer{BayesPref}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
x <- 1 + 1 
}
