% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{calculateLogProb}
\alias{calculateLogProb}
\title{Calculates the log probability of a given guess for weights}
\usage{
calculateLogProb(x, p)
}
\arguments{
\item{x}{Our guess}

\item{p}{A preference elicitation object with the data and observed preferences}
}
\description{
Calculates the log probability of a given guess for weights
}
\examples{
calculateLogProb(c(1, 1.5, 1.0), p)
}

