% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{.calculateLogProb}
\alias{.calculateLogProb}
\title{Calculates the log probability of seeing a given set of preferences}
\usage{
.calculateLogProb(x, p)
}
\arguments{
\item{x}{A guess for our weight vector}

\item{p}{An object of the Bayes preference class}
}
\value{
A scalar log-likelihood of the guess x
}
\description{
Calculates the log probability of seeing a given set of preferences
}
