% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggester.R
\name{.estimateEntropy}
\alias{.estimateEntropy}
\title{Calculates the expected posterior entropy of the prefel object if x and y are compared. Ignores 
the odds of indifference preferences, as using them would increase runtime 50\% without much gain.}
\usage{
.estimateEntropy(p, currentGuess, x, y)
}
\arguments{
\item{p}{An object of class BayesPrefClass.}

\item{currentGuess}{The current best estimate for our weight vector.}

\item{x}{Possible comparison 1}

\item{y}{Possible comparison 2}
}
\description{
Calculates the expected posterior entropy of the prefel object if x and y are compared. Ignores 
the odds of indifference preferences, as using them would increase runtime 50\% without much gain.
}
