% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{.getLogIndifProb}
\alias{.getLogIndifProb}
\title{Evaluates the likelihood of the observed indifference preferences}
\usage{
.getLogIndifProb(x, pref, p)
}
\arguments{
\item{x}{the underlying data}

\item{pref}{the stated preference}

\item{p}{the preference elication object}
}
\description{
Evaluates the likelihood of the observed indifference preferences
}
