% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggester.R
\name{.sampleEntropy}
\alias{.sampleEntropy}
\title{Calculates the entropy of a matrix of samples.}
\usage{
.sampleEntropy(X)
}
\arguments{
\item{X}{a matrix where each row is a sample of variables in different columns}
}
\description{
Calculates the entropy of a matrix of samples.
}
