% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{Exp}
\alias{Exp}
\title{A convinience function for generating Exponential priors}
\usage{
Exp(mu)
}
\arguments{
\item{mu}{The mean of the exponential distribution}
}
\value{
A function yielding the PDF a x of a exponential distribution with given statistics.
}
\description{
A convinience function for generating Exponential priors
}
\examples{
Exp(1)(1) == dexp(1,1)
}
\seealso{
Other Priors: \code{\link{Normal}}
}

