% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{\%=\%}
\alias{\%=\%}
\title{A helper function to add in preferences in a user-friendly way.}
\usage{
a \%=\% b
}
\arguments{
\item{a}{The first alternative}

\item{b}{The second alternative}
}
\description{
A helper function to add in preferences in a user-friendly way.
}
\examples{
1 \%=\% 2 # indifferent between 1 and 2
}
\seealso{
Other preferences: 
\code{\link{\%<\%}()},
\code{\link{\%>\%}()}
}
\concept{preferences}
