% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{\%>\%}
\alias{\%>\%}
\title{A helper function to add in preferences in a user-friendly way.}
\usage{
a \%>\% b
}
\arguments{
\item{a}{The preferred row}

\item{b}{The non-preferred row}
}
\description{
A helper function to add in preferences in a user-friendly way.
}
\examples{
1 \%>\% 2 # prefer row 1 to row 2
}
\seealso{
Other preferences: 
\code{\link{\%<\%}()},
\code{\link{\%=\%}()}
}
\concept{preferences}
