% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{Normal}
\alias{Normal}
\title{A convinience function for generating Normal priors}
\usage{
Normal(mu, sigma)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the prior}
}
\value{
A function yielding the PDF a x of a normal distribution with given statistics.
}
\description{
A convinience function for generating Normal priors
}
\examples{
Normal(0, 1)(1) == dnorm(1)
}
\seealso{
Other Priors: \code{\link{Exp}}
}

