% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{prefEl}
\alias{prefEl}
\title{A shortcut to create objects of the class BayesPrefClass}
\usage{
prefEl(...)
}
\arguments{
\item{...}{Arguments to pass on to BayesPrefClass constructor}
}
\description{
A shortcut to create objects of the class BayesPrefClass
}
\examples{
help(BayesPrefClass)
p <- prefEl(data = data.frame(x = c(1,0,1), y = c(0, 1, 1)))
}

