\name{prefeR-package}
\alias{prefeR-package}
\alias{prefeR}
\docType{package}
\title{
\packageTitle{prefeR}
}
\description{
\packageDescription{prefeR}
}
%\details{

%\packageDESCRIPTION{prefeR}
%\packageIndices{prefeR}
%}
\author{
%\packageAuthor{prefeR}

Maintainer: \packageMaintainer{prefeR}
}


\references{
Lepird, J.R., Owen, M.P. & Kochenderfer, M.J. (2015) \emph{
Bayesian Preference Elicitation for Multiobjective Engineering Design Optimization}
Journal of Aerospace Information Systems 12:10, 634-645 
}

%~~ Optionally other standard keywords, one per line, from file %KEYWORDS in the R ~~
%~~ documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
vignette("mtcars")
}
