% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggester.R
\name{suggest}
\alias{suggest}
\title{Suggests a good comparison for the user to make next.}
\usage{
suggest(p, maxComparisons = 10)
}
\arguments{
\item{p}{An object of class BayesPrefClass.}

\item{maxComparisons}{The maximum number of possible comparisons to check. Default: 10.}
}
\value{
A two-element vector of recommended comparisons.
}
\description{
Suggests a good comparison for the user to make next.
}
