\name{BayesPref-package}
\alias{BayesPref-package}
\alias{BayesPref}
\docType{package}
\title{
\packageTitle{BayesPref}
}
\description{
\packageDescription{BayesPref}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesPref}
\packageIndices{BayesPref}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BayesPref}

Maintainer: \packageMaintainer{BayesPref}
}
\references{
Lepird, J.R., Owen, M.P. & Kochenderfer, M.J. (2015) \emph{
Bayesian Preference Elicitation for Multiobjective Engineering Design Optimization}
Journal of Aerospace Information Systems 12:10, 634-645 
}
%}
%~~ Optionally other standard keywords, one per line, from file %KEYWORDS in the R ~~
%~~ documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
x <- 1 + 1 
}
