% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{Exp}
\alias{Exp}
\title{A convinience function for generating Exponential priors}
\usage{
Exp(mu = 1)
}
\arguments{
\item{mu}{The mean of the exponential distribution, i.e. \eqn{1/rate}}
}
\value{
A function yielding the log-PDF at x of a exponential distribution with given statistics.
}
\description{
A convinience function for generating Exponential priors
}
\examples{
Exp(1)(1) == dexp(1,1, log = TRUE)
}
\seealso{
Other priors: \code{\link{Flat}}, \code{\link{Normal}}
}

