% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{\%=\%}
\alias{\%=\%}
\title{A helper function to add in preferences in a user-friendly way}
\usage{
a \%=\% b
}
\arguments{
\item{a}{The nonpreferred row}

\item{b}{The preferred row}
}
\description{
A helper function to add in preferences in a user-friendly way
}
\examples{
1 \%<\% 2 # prefer row 2 to row 1
}
\seealso{
Other Preference.statement.tools: \code{\link{\%<\%}},
  \code{\link{\%>\%}}
}

