% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggester.R
\name{suggest}
\alias{suggest}
\title{Suggests a good comparision for the user to make next}
\usage{
suggest(p)
}
\arguments{
\item{p}{An object of class BayesPrefClass.}
}
\value{
A two-element vector of recommended comparisons.
}
\description{
Suggests a good comparision for the user to make next
}

